# contracts/openapi/platform-api.yaml
openapi: 3.1.0
info:
  title: Anjing AI Platform API
  version: 0.1.0
  summary: V1 control, gateway, billing and operations API contract.
  description: >
    Anjing AI Platform V1 keeps a modular Go service boundary while allowing
    local development through the platform-all command. Business APIs return a
    unified envelope and can be protected by Bearer token or X-API-Key when
    ANJING_AUTH_MODE=enforced. The x-anjing-service-boundaries extension is the
    contract between console entries, owner commands and API groups; it is
    verified against the frontend consoleServiceMap during pnpm verify.
servers:
  - url: http://localhost:18080
    description: platform-all local server
  - url: http://localhost:1820
    description: control-api local server
  - url: http://localhost:1821
    description: gateway-api local server
  - url: http://localhost:1822
    description: billing-service local server
  - url: http://localhost:1823
    description: ops-api local server
security:
  - bearerAuth: []
  - apiKeyAuth: []
...
paths:
...
  /api/control/auth/login:
    post:
      tags: [control]
      operationId: loginControlSession
      summary: Create a local console session.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthLoginRequest"
      responses:
        "200":
          description: Session created.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/AuthSession"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/control/auth/session:
    get:
      tags: [control]
      operationId: getControlSession
      summary: Get the current console session.
      x-anjing-roles: [Administrator, User, Developer, Operator]
      responses:
        "200":
          description: Session returned.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/AuthSessionStatus"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/control/auth/logout:
    post:
      tags: [control]
      operationId: logoutControlSession
      summary: Revoke the current console session.
      x-anjing-roles: [Administrator, User, Developer, Operator]
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthLogoutRequest"
      responses:
        "200":
          description: Session revoked.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/AuthLogoutResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/control/users:
    get:
      tags: [control]
      operationId: listUsers
      summary: List console users.
      x-anjing-roles: [Administrator]
      responses:
        "200":
          description: Users returned.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [control]
      operationId: createUser
      summary: Invite a new console user.
      x-anjing-roles: [Administrator]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User invited.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/control/users/activate:
    post:
      tags: [control]
      operationId: activateUser
      summary: Activate an invited console user.
      x-anjing-roles: [Administrator]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivateUserRequest"
      responses:
        "200":
          description: User activated.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/control/applications:
    get:
      tags: [control]
      operationId: listApplications
      summary: List onboarding applications.
      x-anjing-roles: [Administrator, User, Developer]
      responses:
        "200":
          description: Applications returned.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Application"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [control]
      operationId: createApplication
      summary: Create an onboarding application and default API key reference.
      x-anjing-roles: [Administrator, User, Developer]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApplicationRequest"
      responses:
        "201":
          description: Application created.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Application"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/control/applications/activate:
    post:
      tags: [control]
      operationId: activateApplication
      summary: Mark an onboarding application as active after validation.
      x-anjing-roles: [Administrator, User, Developer]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivateApplicationRequest"
      responses:
        "200":
          description: Application activated.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Application"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/control/applications/rotate-key:
    post:
      tags: [control]
      operationId: rotateApplicationKey
      summary: Rotate the default API key for an onboarding application.
      x-anjing-roles: [Administrator, User, Developer]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RotateApplicationKeyRequest"
      responses:
        "200":
          description: Application API key rotated.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Application"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/control/roles:
    get:
      tags: [control]
      operationId: listRolePolicies
      summary: List role policies.
      x-anjing-roles: [Administrator, Developer]
      responses:
        "200":
          description: Role policies returned.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/RolePolicy"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/control/api-keys:
    get:
      tags: [control]
      operationId: listAPIKeys
      summary: List masked API keys.
      x-anjing-roles: [Administrator, Developer]
      responses:
        "200":
          description: API keys returned.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/APIKey"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/control/api-keys/revoke:
    post:
      tags: [control]
      operationId: revokeAPIKey
      summary: Revoke an API key.
      x-anjing-roles: [Administrator]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevokeAPIKeyRequest"
      responses:
        "200":
          description: API key revoked.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/APIKey"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/control/credentials:
    get:
      tags: [control]
      operationId: listCredentials
      summary: List credential references.
      x-anjing-roles: [Administrator, Developer]
      responses:
        "200":
          description: Credential references returned.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Credential"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/control/credentials/rotate:
    post:
      tags: [control]
      operationId: rotateCredential
      summary: Rotate a credential reference and return the new active ref.
      x-anjing-roles: [Administrator]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RotateCredentialRequest"
      responses:
        "200":
          description: Credential rotated.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Credential"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/gateway/routes:
    get:
      tags: [gateway]
      operationId: listGatewayRoutes
      summary: List API gateway routes.
      x-anjing-roles: [Administrator, Developer]
      responses:
        "200":
          description: Gateway routes returned.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/GatewayRoute"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [gateway]
      operationId: createGatewayRoute
      summary: Create a gateway route draft.
      x-anjing-roles: [Administrator, Developer]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGatewayRouteRequest"
      responses:
        "201":
          description: Gateway route created.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/GatewayRoute"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: demo-token or session-token
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  responses:
...
    BadRequest:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unauthorized:
      description: Missing or invalid credential.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Forbidden:
      description: Credential is valid but the role is not allowed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    SuccessEnvelope:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
        data:
          description: Operation payload.
    ErrorEnvelope:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          const: false
        error:
          $ref: "#/components/schemas/APIError"
    APIError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          examples: [bad_request, unauthorized, forbidden, not_found, internal_error]
        message:
          type: string
    Principal:
      type: object
      required: [subject, role, method]
      properties:
        subject:
          type: string
        role:
          type: string
          enum: [Administrator, User, Developer, Operator]
        method:
          type: string
          enum: [bearer, api_key, permissive, passwordless, session]
    AuthLoginRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
        role:
          type: string
          enum: [Administrator, User, Developer, Operator]
          description: Optional dev profile role when the email is not in the seed user store.
    AuthSession:
      type: object
      required: [token, principal, expiresAt]
      properties:
        token:
          type: string
          description: Signed local session token.
        principal:
          $ref: "#/components/schemas/Principal"
        expiresAt:
          type: string
          format: date-time
    AuthSessionStatus:
      type: object
      required: [authenticated, principal]
      properties:
        authenticated:
          type: boolean
        principal:
          $ref: "#/components/schemas/Principal"
    AuthLogoutRequest:
      type: object
      additionalProperties: false
    AuthLogoutResponse:
      type: object
      required: [revoked]
      properties:
        revoked:
          type: boolean
...
    User:
      type: object
      required: [id, email, org, role, mfa, status, createdAt]
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        org:
          type: string
        role:
          type: string
          enum: [Administrator, User, Developer, Operator]
        mfa:
          type: string
        status:
          type: string
        createdAt:
          type: string
    CreateUserRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
        org:
          type: string
          default: Platform
        role:
          type: string
          enum: [Administrator, User, Developer, Operator]
          default: User
    ActivateUserRequest:
      type: object
      required: [id]
      properties:
        id:
          type: string
    Application:
      type: object
      required: [id, name, owner, environment, apiKey, defaultRoute, plan, status, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        owner:
          type: string
        environment:
          type: string
          enum: [Sandbox, Production]
        apiKey:
          type: string
        defaultRoute:
          type: string
        plan:
          type: string
        status:
          type: string
        createdAt:
          type: string
    CreateApplicationRequest:
      type: object
      required: [name, owner]
      properties:
        name:
          type: string
        owner:
          type: string
          format: email
        environment:
          type: string
          enum: [Sandbox, Production]
          default: Sandbox
        defaultRoute:
          type: string
          default: /api/v1/llm/**
        plan:
          type: string
          default: Free
    ActivateApplicationRequest:
      type: object
      required: [id]
      properties:
        id:
          type: string
    RotateApplicationKeyRequest:
      type: object
      required: [id]
      properties:
        id:
          type: string
    RolePolicy:
      type: object
      required: [id, name, visibleEntries, configScope, restriction, status]
      properties:
        id:
          type: string
        name:
          type: string
        visibleEntries:
          type: string
        configScope:
          type: string
        restriction:
          type: string
        status:
          type: string
    APIKey:
      type: object
      required: [id, name, project, scope, expiresAt, status]
      properties:
        id:
          type: string
        name:
          type: string
        project:
          type: string
        scope:
          type: string
        expiresAt:
          type: string
        status:
          type: string
    RevokeAPIKeyRequest:
      type: object
      required: [id]
      properties:
        id:
          type: string
    Credential:
      type: object
      required: [id, ref, purpose, scope, expiresAt, status, maskedPreview]
      properties:
        id:
          type: string
        ref:
          type: string
        purpose:
          type: string
        scope:
          type: string
        expiresAt:
          type: string
        status:
          type: string
        maskedPreview:
          type: string
    RotateCredentialRequest:
      type: object
      required: [id]
      properties:
        id:
          type: string
    GatewayRoute:
      type: object
      required: [id, route, upstream, auth, limit, status, updatedAt]
      properties:
        id:
          type: string
        route:
          type: string
        upstream:
          type: string
        auth:
          type: string
        limit:
          type: string
        status:
          type: string
        updatedAt:
          type: string
    CreateGatewayRouteRequest:
      type: object
      required: [route]
      properties:
        route:
          type: string
          examples: [/api/v1/llm/**]
        upstream:
          type: string
          default: gateway-api
        limit:
          type: string
          default: 600/min
...